refactor: QuestionRepositoryService, QuestionRepository이 제네릭이 아닌 Question만을 반환하도록 수정
