+++ application/src/main/java/com/comssa/api/question/service/rest/major/implement/AdminMajorMultipleChoiceQuestionUpdateService.java
@@ -1,7 +1,7 @@
 package com.comssa.api.question.service.rest.major.implement;
 
 import com.comssa.persistence.question.domain.major.MajorMultipleChoiceQuestion;
-import com.comssa.persistence.question.service.QuestionRepositoryService;
+import com.comssa.persistence.question.repository.jpa.MajorMultipleChoiceQuestionJpaRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -10,11 +10,11 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 @Transactional
 public class AdminMajorMultipleChoiceQuestionUpdateService {
-	private final QuestionRepositoryService<MajorMultipleChoiceQuestion> questionRepositoryService;
+	private final MajorMultipleChoiceQuestionJpaRepository majorMultipleChoiceQuestionJpaRepository;
 
 	public MajorMultipleChoiceQuestion toggleCanBeShortAnswered(Long questionId) {
-		MajorMultipleChoiceQuestion majorMultipleChoiceQuestion = questionRepositoryService.
-			findById(questionId);
+		MajorMultipleChoiceQuestion majorMultipleChoiceQuestion = majorMultipleChoiceQuestionJpaRepository.
+			findById(questionId).get();
 		majorMultipleChoiceQuestion.toggleCanBeShortAnswered();
 		return majorMultipleChoiceQuestion;
 	}
