+++ core/src/main/kotlin/com/comssa/core/question/service/common/QuestionGetService.kt
@@ -12,7 +12,7 @@ import org.springframework.transaction.annotation.Transactional
 @Service
 @Transactional
 class QuestionGetService(
-	private val questionRepositoryService: QuestionRepositoryService<Question>,
+	private val questionRepositoryService: QuestionRepositoryService,
 ) {
 	fun getQuestionByIdFetchChoiceIfShould(questionId: Long): Question {
 		val question =
