+++ persistence/src/main/java/com/comssa/persistence/question/repository/jpa/MajorMultipleChoiceQuestionJpaRepository.java
@@ -1,4 +1,4 @@
-package com.comssa.persistence.question.repository;
+package com.comssa.persistence.question.repository.jpa;
 
 
 import com.comssa.persistence.question.domain.major.MajorMultipleChoiceQuestion;
