+++ persistence/src/main/java/com/comssa/persistence/question/repository/querydsl/LicenseMultipleChoiceQuestionRepository.java
@@ -53,7 +53,7 @@ public class LicenseMultipleChoiceQuestionRepository
 	) {
 		return getQuery(
 			question,
-			whereCategories(question._super, questionCategories))
+			whereCategoriesAndLevels(question._super, questionCategories, null))
 			.distinct()
 			.leftJoin(question.questionChoices).fetchJoin()
 			.fetch();
